Move components directory under src
