+++ src/components/sections/Connect/Connect.types.ts
